app/requirements.txt
streamlit==1.38.0
langchain_community
langchain_groq
chromadb==0.4.14

pandas
python-dotenv
protobuf==3.20.1
# pysqlite3-binary>=0.5.1
